[core] Make InternalKV configurable in ioctx policy. (#49113)
Defines a InternalKV ioctx in ioctx policy, and assigns it to InternalKV
gRPC service, Manager, and Redis client.

Makes a separate RedisClient specifically for InternalKV on InternalKV's
ioctx, aside from the main io_service which serves all non-InternalKV
Redis tables.

---------

+++ src/ray/common/asio/asio_util.h
@@ -145,7 +145,8 @@ class IOContextProvider {
   instrumented_io_context &GetIOContext() const {
     constexpr int index = Policy::template GetDedicatedIOContextIndex<T>();
     static_assert(
-        index >= -1 && index < Policy::kAllDedicatedIOContextNames.size(),
+        index >= -1 &&
+            index < static_cast<int>(Policy::kAllDedicatedIOContextNames.size()),
         "index out of bound, invalid GetDedicatedIOContextIndex implementation! Index "
         "can only be -1 or within range of kAllDedicatedIOContextNames");
